Use phaser-ton to connect wallet.
