+++ src/consts.ts
@@ -1,0 +1,12 @@
+export const PIPES_AVAILABLE = ['pipe-green', 'pipe-red'];
+export const PIPES_COSTS = [0, 1];
+export const ENDPOINT = 'http://localhost:3000';
+// const ENDPOINT = 'https://flappy.krigga.dev';
+export const TOKEN_RECIPIENT = 'EQBb7bFnXnKAN1DNO3GPKLXPNiEyi4U6-805Y-aBkgJtK_lJ';
+export const TOKEN_MASTER = 'EQBcRUiCkgdfnbnKKYhnPXkNi9BXkq_5uLGRuvnwwaZzelit';
+export const NETWORK = 'testnet';
+export const APP_URL = 'https://t.me/flappybirddevbot/flappybirddev'
+export const APP_MANIFEST_URL = 'https://raw.githubusercontent.com/ton-defi-org/tonconnect-manifest-temp/main/tonconnect-manifest.json';
+
+export const SHOP_RELOAD_INTERVAL = 10000;
+export const BALANCE_RELOAD_INTERVAL = 10000;
