+++ src/phaser-ton/README.md
@@ -4,6 +4,8 @@ TON bindings and utilities for Phaser.io
 
 ## Usage
 
+### Button component
+
 1. Import canvas implementation of `Connect Wallet` button:
 ```typescript
 import { ConnectTelegramWalletButton } from './phaser-ton';
@@ -23,23 +25,34 @@ export class UiScene extends Phaser.Scene {
             // button options
             {
                 style: 'light',
-                appManifestUrl: 'your app manifest url here',
+                manifestUrl: 'your app manifest url here',
             }
         );
     }
 }
 ```
-3. After user connected his wallet you can use it across your app by importing `connectedWallet` object:
+3. After user connected his wallet you can use it across your app by importing `getConnectedWallet` object:
 ```typescript
-import { connectedWallet } from './phaser-ton';
+import { getConnectedWallet } from './phaser-ton';
+const myWallet = getConnectedWallet();
 ```
 Read [Wallet interface documentation](https://ton-connect.github.io/sdk/interfaces/_tonconnect_sdk.Wallet.html) to learn available props and methods.
+
+### Check wallet connection
+
+You can check if user has a connected wallet and use it without rendering the button:
+```typescript
+import { restoreWalletConnection } from './phaser-ton';
+const wallet = await restoreWalletConnection({
+    // ton connection options
+});
+```
 
 ## ConnectTelegramWalletButton options
 
 | Option | Description | Default |
 | -------- | -------- | -------- |
-| appManifestUrl | URL of your app manifest. Read about the [parameters of the manifest](https://docs.ton.org/develop/dapps/ton-connect/protocol/requests-responses#app-manifest) in related docs. | Looks for `tonconnect-manifest.json` file public root of your app, like `https:example.com/tonconnect-manifest.json` |
+| tonParams | Check [@tonconnect/sdk](https://ton-connect.github.io/sdk/interfaces/_tonconnect_sdk.TonConnectOptions.html). The most important thing is to setup [app manifest](https://docs.ton.org/develop/dapps/ton-connect/protocol/requests-responses#app-manifest). | `undefined` |
 | lang | Language of the button. Languages `en` and `ru` supported so far | `en` |
 | style | Button style. Can be `light` or `dark` | `light` |
 | onWalletChange | Callback function that will be called when user connects or disconnects his wallet | `undefined` |
